+++ CHANGELOG.md
@@ -6,7 +6,14 @@ nd this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## Removed
+
 - Remove `SchemaBuilder::override_name` method. [#437](https://github.com/async-graphql/async-graphql/issues/437)
+  
+## Added
+
+- Add `name` and `visible` attributes for `Newtype` macro for define a new scalar. [#437](https://github.com/async-graphql/async-graphql/issues/437)
+- `NewType` macro now also implements `From<InnerType>` and `Into<InnerType>`.
 
 ## [2.7.1]